+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/XLaterTests.java
@@ -16,12 +16,14 @@
 
 package de.gematik.test.tiger.playwright.workflowui;
 
-import static org.assertj.core.api.Assertions.assertThat;
+
+import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import com.microsoft.playwright.Page;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
@@ -38,43 +40,38 @@ class XLaterTests extends AbstractTests {
     page.querySelector("#test-execution-pane-tab").click();
     page.locator("#test-webui-slider").click();
     assertAll(
-        () -> assertThat(page.locator("#rbellog_details_pane").isVisible()).isTrue(),
+        () -> assertThat(page.locator("#rbellog_details_pane")).isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#webui-navbar")
-                        .isVisible())
-                .isTrue(),
+                        ).isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#dropdown-hide-button")
-                        .isVisible())
-                .isTrue(),
+                        ).isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#filterModalBtn")
-                        .isVisible())
-                .isTrue(),
+                        ).isVisible(),
         () ->
             assertThat(
-                    page.frameLocator("#rbellog-details-iframe").locator("#exportMsgs").isVisible())
-                .isTrue(),
+                    page.frameLocator("#rbellog-details-iframe").locator("#exportMsgs")).isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#dropdown-page-selection")
-                        .isVisible())
-                .isTrue(),
+                        ).isVisible(),
         () ->
-            assertThat(
+            Assertions.assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#pageNumberDisplay")
                         .textContent())
                 .endsWith("1"),
         () ->
-            assertThat(
+            Assertions.assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#pageSizeDisplay")
                         .textContent())
@@ -83,28 +80,23 @@ class XLaterTests extends AbstractTests {
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#dropdown-page-size")
-                        .isVisible())
-                .isTrue(),
+                        ).isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#routeModalBtn")
-                        .isVisible())
-                .isFalse(),
+                        ).not().isVisible(),
         () ->
             assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#scrollLockBtn")
-                        .isVisible())
-                .isFalse(),
+                        ).not().isVisible(),
         () ->
             assertThat(
-                    page.frameLocator("#rbellog-details-iframe").locator("#resetMsgs").isVisible())
-                .isFalse(),
+                    page.frameLocator("#rbellog-details-iframe").locator("#resetMsgs")).not().isVisible(),
         () ->
             assertThat(
-                    page.frameLocator("#rbellog-details-iframe").locator("#importMsgs").isVisible())
-                .isFalse());
+                    page.frameLocator("#rbellog-details-iframe").locator("#importMsgs")).not().isVisible());
   }
 
   @Test
@@ -113,19 +105,19 @@ class XLaterTests extends AbstractTests {
     page.locator("#test-webui-slider").click();
     assertAll(
         () ->
-            assertThat(
+            Assertions.assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#test-rbel-section .test-card")
                         .count())
                 .isPositive(),
         () ->
-            assertThat(
+            Assertions.assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#test-rbel-section .test-card-header")
                         .count())
                 .isPositive(),
         () ->
-            assertThat(
+            Assertions.assertThat(
                     page.frameLocator("#rbellog-details-iframe")
                         .locator("#test-rbel-section .test-card-content")
                         .count())
@@ -141,15 +133,15 @@ class XLaterTests extends AbstractTests {
     await().untilAsserted(() -> assertNotNull(externalPage.locator("#routeModalBtn")));
     externalPage.locator("#routeModalBtn").click();
     assertAll(
-        () -> assertThat(externalPage.locator("#routeModalDialog").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#routingModalButtonClose").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#addNewRouteBtn").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#addNewRouteFromField").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#addNewRouteFromField").textContent()).isEmpty(),
-        () -> assertThat(externalPage.locator("#addNewRouteToField").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#addNewRouteToField").textContent()).isEmpty());
+        () -> assertThat(externalPage.locator("#routeModalDialog")).isVisible(),
+        () -> assertThat(externalPage.locator("#routingModalButtonClose")).isVisible(),
+        () -> assertThat(externalPage.locator("#addNewRouteBtn")).isVisible(),
+        () -> assertThat(externalPage.locator("#addNewRouteFromField")).isVisible(),
+        () -> assertThat(externalPage.locator("#addNewRouteFromField")).isEmpty(),
+        () -> assertThat(externalPage.locator("#addNewRouteToField")).isVisible(),
+        () -> assertThat(externalPage.locator("#addNewRouteToField")).isEmpty());
     externalPage.locator("#routingModalButtonClose").click();
-    assertThat(externalPage.locator("#routeModalDialog").isVisible()).isFalse();
+    assertThat(externalPage.locator("#routeModalDialog")).not().isVisible();
     externalPage.close();
   }
 
@@ -161,16 +153,16 @@ class XLaterTests extends AbstractTests {
     Page externalPage = page.waitForPopup(() -> page.locator("#test-rbel-webui-url").click());
     await().untilAsserted(() -> assertNotNull(externalPage.locator("#routeModalBtn")));
     assertAll(
-        () -> assertThat(externalPage.locator("#test-tiger-logo").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#routeModalBtn").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#scrollLockBtn").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#dropdown-hide-button").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#filterModalBtn").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#resetMsgs").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#exportMsgs").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#dropdown-page-selection").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#dropdown-page-size").isVisible()).isTrue(),
-        () -> assertThat(externalPage.locator("#importMsgs").isVisible()).isTrue());
+        () -> assertThat(externalPage.locator("#test-tiger-logo")).isVisible(),
+        () -> assertThat(externalPage.locator("#routeModalBtn")).isVisible(),
+        () -> assertThat(externalPage.locator("#scrollLockBtn")).isVisible(),
+        () -> assertThat(externalPage.locator("#dropdown-hide-button")).isVisible(),
+        () -> assertThat(externalPage.locator("#filterModalBtn")).isVisible(),
+        () -> assertThat(externalPage.locator("#resetMsgs")).isVisible(),
+        () -> assertThat(externalPage.locator("#exportMsgs")).isVisible(),
+        () -> assertThat(externalPage.locator("#dropdown-page-selection")).isVisible(),
+        () -> assertThat(externalPage.locator("#dropdown-page-size")).isVisible(),
+        () -> assertThat(externalPage.locator("#importMsgs")).isVisible());
     externalPage.close();
   }
 
@@ -178,35 +170,24 @@ class XLaterTests extends AbstractTests {
   void testCServers() {
     page.querySelector("#test-tiger-logo").click();
     assertAll(
-        () -> assertThat(page.locator("#test-sidebar-server-status-box").isVisible()).isTrue(),
+        () -> assertThat(page.locator("#test-sidebar-server-status-box")).isVisible(),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-serverbox").count())
-                .isEqualTo(3),
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-serverbox")).hasCount(3),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-name")
-                        .count())
-                .isEqualTo(3),
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-name")).hasCount(3),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-status")
-                        .count())
-                .isEqualTo(3),
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-status")).hasCount(3),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-url")
-                        .count())
-                .isEqualTo(2),
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-url")).hasCount(2),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-url-icon")
-                        .count())
-                .isEqualTo(2),
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-url-icon")).hasCount(2),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-log-icon")
-                        .count())
-                .isEqualTo(3));
+                    page.locator("#test-sidebar-server-status-box .test-sidebar-server-log-icon")).hasCount(3));
   }
 }